[Feature] refactor template, filter function, add form enable & disable elements

+++ src/ajax-form-submit-success-handler.js
@@ -1,15 +1,16 @@
+import { OBJECT, FUNCTION, STRING_NON_BLANK } from './js-constant.js'
 import {
   assert,
   formatUrl,
-  startsWith,
   isTrue,
   isArray,
   isObject,
   isFunction,
   isNotBlank,
   hasValue,
+  objectKeys,
+  objectEntries,
   valueToString,
-  stringToValue,
   findObjectValue,
   addBasePath
 } from './js-utils.js'
@@ -27,31 +28,27 @@ import { createProperty } from './js-property-factory.js'
 
 const CLASS_NAME = 'ajax-form-submit-success'
 const SKELETON_CLASS_NAME = `${CLASS_NAME}-skeleton`
+const HANDLERS = {}
 const LIFECYCLES = [
-  { name: 'before', required: false },
-  { name: 'validation', required: false },
-  { name: 'request', required: false },
-  { name: 'response', required: false },
+  { name: 'before' },
+  { name: 'validation' },
+  { name: 'request' },
+  { name: 'response' },
   { name: 'after', required: true },
-  { name: 'error', required: false },
+  { name: 'error' },
 ]
 
-let HANDLERS = {
-  redirect: createHandler(handleRedirect),
-  show: createHandler(handleShow),
-  hide: createHandler(handleHide),
-  querystring: createHandler(handleUpdateQueryString),
-  storage: createHandler(handleStorage),
-  display: createHandler(handleDisplay()),
-  event: createHandler(handleEvent()),
-}
+addHandler('redirect', handleRedirect)
+addHandler('show', handleShow)
+addHandler('hide', handleHide)
+addHandler('querystring', handleUpdateQueryString)
+addHandler('storage', handleStorage)
+addHandler('display', handleDisplay())
+addHandler('event', handleEvent())
 
 export default class AjaxFormSubmitSuccessHandler {
 
-  static add = (type, callback) => {
-    assert(isNotBlank(type), 1, 'NonBlankString')
-    HANDLERS[type] = createHandler(callback)
-  }
+  static add = addHandler
 
   #root
   #datasetHelper
@@ -64,7 +61,6 @@ export default class AjaxFormSubmitSuccessHandler {
 
   constructor(opt = {}) {
     const { handlerProps, attrKey, ...payload } = opt
-
     this.#root = payload.root
     this.#datasetHelper = payload.datasetHelper
     this.#defaultHandlerProps = handlerProps
@@ -72,10 +68,10 @@ export default class AjaxFormSubmitSuccessHandler {
     this.#attrInputName = this.#datasetHelper.keyToInputName(attrKey)
     this.#attrRegex = new RegExp(String.raw`_?${this.#attrInputName}-`, 'g')
     this.#payload = payload
-    this.#registerLifecycle() 
+    this.#registerLifecycleMethod() 
   }
 
-  #registerLifecycle() {
+  #registerLifecycleMethod() {
     for (const lifecycle of LIFECYCLES) {
       this[lifecycle.name] = (opts, input, output, type) => this.#run(lifecycle, opts, input, output, type)
     }
@@ -83,34 +79,28 @@ export default class AjaxFormSubmitSuccessHandler {
 
   #run(lifecycle, opts, input, output, type) {
     this.#updateHandlerProps(opts)
-    const types = Object.keys(this.#handlerProps)
+    const types = objectKeys(this.#handlerProps)
     if (isNotBlank(type) && !types.includes(type))
       return
 
     const selectTypes = type ? [ type ] : types
-    selectTypes.forEach(selectType => {
+    for (const selectType of selectTypes) {
       const handler = HANDLERS[selectType]?.[lifecycle.name]
-      if (lifecycle.required) {
-        assert(isFunction(handler), `Could not find "${selectType}" in successHandlers`)
-      }
-      this.#handlerProps[selectType].forEach(prop => handler?.(input, output, prop, this.#payload))
-    })
+      lifecycle.required && assert(isFunction(handler), `Could not find "${selectType}" in successHandlers`)
+      this.#handlerProps[selectType]
+        .forEach(prop => handler?.(input, output, prop, { ...this.#payload, ...opts }))
+    }
   }
 
   #updateHandlerProps(handlerProps) {
     this.#handlerProps = {}
-    this.#defaultHandlerProps
-    if (isObject(this.#defaultHandlerProps)) {
-      for (const [key, value] of Object.entries(this.#defaultHandlerProps)) {
-        this.#handlerProps[key] = createProperty(value)
-      }
+    for (const [key, value] of objectEntries(this.#defaultHandlerProps)) {
+      this.#handlerProps[key] = createProperty(value)
     }
 
-    if (isObject(handlerProps)) {
-      for (const [key, value] of Object.entries(handlerProps)) {
-        if (key.includes(this.#attrInputName))
-          this.#handlerProps[key.replace(this.#attrRegex, '')] = createProperty(value)
-      }
+    for (const [key, value] of objectEntries(handlerProps)) {
+      if (key.includes(this.#attrInputName))
+        this.#handlerProps[key.replace(this.#attrRegex, '')] = createProperty(value)
     }
 
     if (isElement(this.#root)) {
@@ -123,19 +113,18 @@ export default class AjaxFormSubmitSuccessHandler {
   }
 }
 
-function createHandler(callback) {
+function addHandler(type, callback) {
+  assert(isNotBlank(type), 1, STRING_NON_BLANK)
   if (isFunction(callback)) {
-    return { after: callback }
+    HANDLERS[type] = { after: callback }
   } else if (isObject(callback)) {
-    let result = {}
-    for (const { name, required } of LIFECYCLES) {
-      isFunction(callback[name]) && (result[name] = callback[name])
-      if (required)
-        assert(isFunction(result[name]), `handler.${name} must be Function`)
-    }
-    return result
+    HANDLERS[type] = LIFECYCLES.reduce((acc, { name, required }) => {
+      isFunction(callback[name]) && (acc[name] = callback[name])
+      required && assert(hasValue(acc[name]), `handler.${name} must be Function`)
+      return acc
+    }, {})
   } else {
-    assert(false, 1, 'Function or Object')
+    assert(false, 1, [OBJECT, FUNCTION])
   }
 }
 
@@ -145,9 +134,8 @@ export function handleEvent(defaultEventName) {
     if (isNotBlank(defaultEventName) && events.size === 0)
       events.add(defaultEventName)
 
-    const payload = { input, output, props }
-    getTargets(target, root).forEach(elem =>
-      events.forEach(eventName => triggerEvent(elem, eventName, payload)))
+    getTargets(target, root).forEach(elem => events.forEach(eventName =>
+      triggerEvent(elem, eventName, { input, output, props })))
   }
 }
 
@@ -168,44 +156,44 @@ function handleRedirect(input, output, { target, type, param }, { basePath }) {
     default:
       if (!isNotBlank(url)) {
         location.reload()
-        break
+      } else {
+        const outputObj = isObject(output) ? output : { value: output }
+        url = addBasePath(formatUrl(formatUrl(target[0], input), outputObj), basePath)
+        let params = new URLSearchParams()
+        param?.forEach?.(key => {
+          const inputValue = findObjectValue(input, key)
+          const outputValue = findObjectValue(outputObj, key)
+          inputValue.exist && params.set(key, inputValue.value)
+          outputValue.exist && params.set(key, outputValue.value)
+        })
+
+        if (params.size > 0)
+          url += `?${params.toString()}`
+
+        location.href = url
       }
-      const outputObj = isObject(output) ? output : { value: output }
-      url = addBasePath(formatUrl(formatUrl(target[0], input), outputObj), basePath)
-      let params = new URLSearchParams()
-      param?.forEach?.(key => {
-        const inputValue = findObjectValue(input, key)
-        const outputValue = findObjectValue(outputObj, key)
-        inputValue.exist && params.set(key, inputValue.value)
-        outputValue.exist && params.set(key, outputValue.value)
-      })
-
-      if (params.size > 0)
-        url += `?${params.toString()}`
-
-      location.href = url
   }
 }
 
 function handleDisplay() {
+  const group = 'skeleton'
   return {
-    before: (input, output, { target }, { domHelper, datasetHelper }) => {
-      getTargets(target)
-        .filter(elem => !isTrue(datasetHelper.getValue(elem, 'append')))
-        .forEach(elem => domHelper?.clearElement?.(elem))
+    before: (input, output, { target }, { append, domHelper, datasetHelper }) => {
+      getTargets(target).forEach(elem => {
+        const props = createProperty(datasetHelper.getValue(elem, 'template'))[0]
+        !isTrue(append) && !isTrue(props.append?.[0]) && domHelper?.clearElement?.(elem)
+      })
     },
-    request: (input, output, { target, skeleton }, { domHelper }) => {
-      if (!skeleton?.[0] || !input.size)
-        return
-
-      const mockOutput = Array.from({ length: input.size }, () => ({}))
-      getTargets(target).forEach(elem =>
-        domHelper?.setValueToElement?.(elem, mockOutput, { template: skeleton[0] }))
-      return true
+    request: (input, output, { target }, { domHelper, datasetHelper }) => {
+      const mock = Array.from({ length: input.size || 1 }, () => ({}))
+      getTargets(target).forEach(elem => {
+        const { skeleton: [template] = [] } = createProperty(datasetHelper.getValue(elem, 'template'))[0]
+        isNotBlank(template) && domHelper?.setValueToElement?.(elem, mock, { template, group })
+      })
     },
     after: (input, output, { target }, { domHelper }) => {
       getTargets(target).forEach(elem => {
-        querySelector(`.${SKELETON_CLASS_NAME}`, elem).forEach(skeleton => skeleton.remove())
+        domHelper?.clearElement?.(elem, group)
         domHelper?.setValueToElement?.(elem, output)
       })
     }
@@ -220,33 +208,35 @@ function handleHide(input, output, { target }, { root }) {
   hideElements(getTargets(target, root))
 }
 
-function handleUpdateQueryString(input, output, { target }) {
-  if (!URL || !location || !history || !history.pushState)
-    return
-
+function handleUpdateQueryString(input, output, { add, remove, value }) {
   const { host, protocol, pathname } = location
-  const includes = []
-  const excludes = []
+  const includes = new Set([ ...(add || []), ...(value || []) ])
+  const excludes = new Set(remove || [])
   const url = new URL(`${protocol}//${host}${pathname}`)
-  
-  target.forEach(str => {
-    const { exist, value } = startsWith(str, '-')
-    exist ? excludes.push(value) : includes.push(value)
-  })
+  const searchParams = url.searchParams
+  const hasIncludes = includes.size > 0
+  const hasExcludes = !hasIncludes && excludes?.size > 0
 
-  if (includes.length > 0) {
-    for(const key in input){
-      if (includes.includes(key)) {
-        setQueryString(url, key, input[key])
+  objectKeys(input).forEach(key => {
+    if (hasIncludes ? includes.has(key) : hasExcludes ? !excludes.has(key) : true) {
+      const val = input[key]
+      if (isArray(val)) {
+        searchParams.delete(key)
+        val.forEach(item => searchParams.append(key, item))
+      } else if (isObject(val)) {
+        searchParams.set(key, JSON.stringify(val))
+      } else if (hasValue(val)) {
+        if (typeof val === 'string') {
+          isNotBlank(val) ? searchParams.set(key, val) : searchParams.delete(key)
+        } else {
+          searchParams.set(key, val)
+        }
+      } else {
+        searchParams.delete(key)
       }
     }
-  } else {
-    for(const key in input){
-      if (!excludes.includes(key)) {
-        setQueryString(url, key, input[key])
-      }
-    }
-  }
+  })
+  
   history.replaceState({ path: url.href }, '', url.href)
 }
 
@@ -264,20 +254,8 @@ function handleStorage(input, output, { value }, { root, prefix }) {
     if (key === 'timestamp') {
       localStorage.setItem(storageKey, Date.now())
     } else {
-      const value = findObjectValue(data, key).value
-        if (hasValue(value) && isNotBlank(key))
-          localStorage.setItem(storageKey, valueToString(value))
+      const { exist, value } = findObjectValue(data, key)
+        exist && localStorage.setItem(storageKey, valueToString(value))
     }
   })
-}
-
-function setQueryString(url, key, value) {
-  if (isArray(value)) {
-    url.searchParams.delete(key)
-    value.forEach(item => url.searchParams.append(key, item))
-  } else if (isObject(value)) {
-    url.searchParams.set(key, JSON.stringify(value))
-  } else if (hasValue(value)) {
-    url.searchParams.set(key, value)
-  }
 }
